add get_many() function, make update() efficient
Adds a new service.get_many() function to the library that takes a series of
filter parameters and returns service objects matching all the filters.

Also changes the service.update() function to be much more efficient, only
actually making changes to the indexes when needed and the primary service
record is updated only on changes to fields.